push

+++ database/migrations/2022_11_07_142549_create_sponsorizations_table.php
@@ -17,7 +17,7 @@ class CreateSponsorizationsTable extends Migration
             $table->id();
 
             $table->string('name');
-            $table->double('price', 2, 2);
+            $table->decimal('price', 4, 2);
             $table->tinyInteger('duration')->unsigned();
             
             $table->timestamps();
